Move fingerprint types into fingerprint module
